+++ tests/test_calibration.py
@@ -1,6 +1,6 @@
 import os
 import re
-from datetime import datetime, timedelta
+from datetime import datetime, timedelta, timezone
 from pathlib import Path
 
 import numpy as np
@@ -17,6 +17,9 @@ from extra.calibration import (
     JUNGFRAUConditions,
     LPDConditions,
     SingleConstant,
+    DetectorData,
+    DetectorModule,
+    SourceNameFormatter
 )
 
 # Most of these tests use saved HTTP responses by default (with pytest-recording).
@@ -319,6 +322,12 @@ def test_AGIPD_from_correction():
     # available for lazy loading.
     assert agipd_cd["Offset", "AGIPD00"]._metadata['report_id'] == 6512
 
+    # Smoketest
+    report_tbl = agipd_cd.reports_info()
+    md = report_tbl._repr_markdown_()
+    assert '[6512](https://in.xfel.eu/calibration/reports/6512)' in md
+    report_tbl._repr_latex_()
+
 
 @pytest.mark.vcr
 def test_LPD_from_correction():
@@ -377,3 +386,98 @@ def test_conditions_markdown():
     cond = LPDConditions()
     md = MarkdownFormatter()(cond)  # Smoketest
     assert isinstance(md, str)
+
+
+pdu_date_kw = dict(pdu_snapshot_at=datetime(
+    year=2025, month=10, day=27, hour=15, minute=50, second=13,
+    tzinfo=timezone(timedelta(hours=1))))
+
+
+@pytest.mark.vcr
+def test_DetectorData_from_identifier():
+    # SPB-AGIPD, multi-module detector with full CalCat entries
+    agipd = DetectorData.from_identifier('SPB_DET_AGIPD1M-1', **pdu_date_kw)
+    repr(agipd)
+
+    assert agipd
+    assert agipd.identifier == 'SPB_DET_AGIPD1M-1'
+    assert agipd.detector_type == 'AGIPD-Type'
+    assert len(agipd) == agipd.number_of_modules
+    assert list(agipd)[0] == next(iter(agipd.keys())) == 'AGIPD00'
+    assert isinstance(next(iter(agipd.values())), DetectorModule)
+    assert agipd[0] == agipd['AGIPD00']
+    assert agipd.source_names == [
+        f'SPB_DET_AGIPD1M-1/DET/{i}CH0:xtdf'
+        for i in range(len(agipd.source_names))]
+
+    # PDU
+    pdu = agipd[0]
+    assert pdu.aggregator == 'AGIPD00'
+    assert pdu.ccv_params == (
+        'AGIPD_SIV1_AGIPDV11_M517', 101003000000, 'AGIPD-Type')
+    assert pdu.source_name == 'SPB_DET_AGIPD1M-1/DET/0CH0:xtdf'
+
+    # FXE-JFHZ, single-module detector with partial CalCat entries
+    jfhz = DetectorData.from_identifier('FXE_XAD_JFHZ', **pdu_date_kw)
+    repr(jfhz)
+
+    assert jfhz
+    assert len(jfhz) == 1
+    assert jfhz.number_of_modules is None
+
+    with pytest.raises(AssertionError):
+        jfhz.first_module_index
+
+    with pytest.raises(AssertionError):
+        jfhz.source_names
+
+    # SQS-DSSC, PDU-less detector
+    dssc = DetectorData.from_identifier('SQS_DET_DSSC1M-1', **pdu_date_kw)
+    repr(dssc)
+
+    assert not dssc
+    assert len(dssc) == 0
+
+    with pytest.raises(ValueError):
+        dssc.detector_type
+
+
+@pytest.mark.vcr
+def test_DetectorData_from_instrument():
+    with pytest.raises(ValueError):
+        DetectorData.from_instrument('SPB')  # More than one detector.
+
+    with pytest.raises(ValueError):
+        DetectorData.from_instrument('SPB', '*JF*')  # More than one JF.
+
+    # Sufficiently narrow glob.
+    jf4m = DetectorData.from_instrument('SPB', '*JF4M', **pdu_date_kw)
+    assert jf4m.identifier == 'SPB_IRDA_JF4M'
+
+    # Instrument with single detector.
+    hirex = DetectorData.from_instrument('SA1', **pdu_date_kw)
+    assert hirex.identifier == 'SA1_XTD9_HIREX'
+
+
+@pytest.mark.vcr
+def test_DetectorData_list_by_instrument():
+    assert DetectorData.list_by_instrument('SCS') == [
+        'SCS_DET_DSSC1M-1', 'SCS_HRIXS_JUNGF', 'SCS_XOX_GH21',
+        'SCS_XOX_GH22', 'SCS_DET_DSSC2']
+
+
+@pytest.mark.vcr
+def test_DetectorData_from_CalibrationData():
+    agipd_cd = CalibrationData.from_report(3757)
+    assert agipd_cd.detector.identifier == 'SPB_DET_AGIPD1M-1'
+
+
+def test_DetectorData_SourceNameFormatter():
+    fmt = SourceNameFormatter()
+
+    assert 'SPB_DET_AGIPD1M-1/DET/4CH0:xtdf' == \
+        fmt.format('SPB_DET_AGIPD1M-1/DET/{modno}CH0:xtdf', modno=4)
+    assert 'SPB_IRDA_JF4M/DET/JNGFR03:daqOutput' == \
+        fmt.format('SPB_IRDA_JF4M/DET/JNGFR{modno:02d}:daqOutput', modno=3)
+    assert 'HED_TST_AGIPDHZ3/DET/84CH0:xtdf' == \
+        fmt.format('HED_TST_AGIPDHZ3/DET/{modno+83}CH0:xtdf', modno=1)