+++ docs/changelog.md
@@ -19,6 +19,8 @@
 Added:
 
 - [AGIPDConditions][extra.calibration.AGIPDConditions] now recognizes current source constant types
+- [DetectorData][extra.calibration.DetectorData] to obtain detector metadata and
+  module mapping from the Calibration Cataloge (!408).
 - [SpectrometerCalibration][extra.gui.jupyter.SpectrometerCalibration] to
   provide a Jupyter widget for energy calibration of 2D X-ray spectrum data
   (!363).
@@ -26,6 +28,9 @@ Added:
   supports `DataCollection` objects to reference a point in time
 - [CalibrationData.from_correction][extra.calibration.CalibrationData.from_correction]
   method to find the constants used to correct a particular run.
+- [CalibrationData.reports_info][extra.calibration.CalibrationData.reports_info]
+  method to show a summary of the reports associated with groups of calibration
+  constants.
 - Pre-built packages will be available on PyPI for Python 3.10 - 3.13 from the
   next release, rather than only Python 3.10 (!377).
 - [Scan.group_data()][extra.components.Scan.group_data] method to make an xarray